esp32-audio-board: refactor PCB
Added many changes to the PCB, which hopefully
will protect the PCB from EMI even better and
will ensure that all signals have a short return
path to GND plane. Added pads for a boot button
to allow the board to be flashed even if the
auto-download circuit will not work.

Signed-off-by: RaresCon <rares.constantin2002@gmail.com>