esp32-audio-board: fix VI for audio pwr
The audio 3V3 was coming directly from the battery, which
may not be present and the VBUS should power the audio
circuits. The production board has this problem, but it
should not be affected if a battery is plugged in, as
the battery voltage will be present and the battery will
power the audio circuits correctly even when charging.

The charging led is now hooked correctly to the charger
IC on the STAT pin.

Signed-off-by: RaresCon <rares.constantin2002@gmail.com>